requirements.txt
APScheduler==3.6.3
bcrypt==4.0.1
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==41.0.1
flake8==3.9.2
flake8-docstrings==1.6.0
idna==3.4
mccabe==0.6.1
packaging==23.1
paramiko==3.2.0
Pillow==9.5.0
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.3.1
PyNaCl==1.5.0
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2023.3
pyzabbix==1.3.0
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
